requirements/dev-requirements.txt

autopep8==1.2.4
beautifulsoup4==4.6.0
django-debug-toolbar==1.5
ipdb==0.10.1
pdbpp==0.9.2
pip-review==0.4
pipdeptree==0.10.1
pygraphviz==1.3.1